automations/multi-page-web-scraper/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
gspread>=6.0.0
google-auth>=2.0.0
